requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-bootstrap-v5==1.0.11
django-js-asset==2.0.0
django-object-tools==2.0.0
django-taggit==4.0.0
django-treebeard==4.7
idna==3.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
